fix(office-transform): refuse data-table ranges and docx break characters
ST_CellFormulaType has four values. The script enumerated three: shared and array were
refused, dataTable fell through to the ordinary-formula path and was written silently. A
data table stores its grid the way an array formula does — one master carrying the `<f>`
that owns the range, members holding a bare `<v>` — so writing the master deletes the only
`<f>` defining the whole grid and leaves the rest as orphan literals. Verified on a
workbook holding all four kinds: before, B4 (dataTable master) and B5 (member) were both
accepted, `<f t="dataTable" ref="B4:B6">` disappeared and B5/B6 became loose numbers, while
array and shared were refused as designed. A write naming one cell quietly changed three,
which is what reject_shared_formula's own docstring exists to prevent.

Collecting the range up front closes both the master and the member case at once, because a
master sits inside its own `ref`. The refusal carries the kind so it says "data table" —
naming it "array formula" would send the caller looking for the wrong thing. Renamed
array_formula_ranges to grouped_formula_ranges, since the old name stopped being true. The
module docstring pointed at a `reject_grouped_formula` that does not exist; it is in the
sentence being corrected anyway, so it now names the two functions that do the work.

docx replacement text carrying a tab or newline was written into one `<w:t>`, where
WordprocessingML expects `<w:tab/>` and `<w:br/>` elements. Refuse rather than translate:
the mapping is not reversible, since `w:br`, `w:cr` and a page break all read back as the
same newline, so translating would be guessing which one the caller meant. This is
reachable through the skill's own round trip and silent there — extracting a paragraph
holding a real `w:tab` gives you a tab character, and writing the edited string back drops
the element while the text still reads the same. Only docx goes through the new check;
reject_invalid_xml_text is untouched, because a newline in an xlsx inline string
legitimately means a line break inside the cell. Verified both, including that xlsx still
accepts one.

Two nested anchor fields were read with `.get()` without checking they were objects, on the
path SKILL.md explicitly invites the caller onto ("build the anchor JSON yourself"): a
replacements element, and a pptx `tableCell` written as "A1" rather than {row, col}. Both
raised AttributeError with a traceback, which reads to the calling agent as a broken script
rather than a bad argument. The top-level `--anchor null` case the review also named is
left alone: no caller reaches it, and the file already declines to guard impossible input.

mtimeMs now says `.int()`, matching the whole-millisecond invariant the freshness check
documents — its 2 ms window is only sound because both sides floor the same nanosecond
timestamp. Node's raw stats.mtimeMs is fractional; the main-side stat helpers already floor
it, so this states an invariant that already holds rather than changing behaviour.

+++ src/renderer/types/selectionReference.ts
@@ -105,7 +105,9 @@ export type DocumentAnchor = z.infer<typeof DocumentAnchorSchema>
 /** Snapshot of the source file's identity at capture time, checked before use. */
 const SelectionFileStampSchema = z.object({
   size: z.number().int().nonnegative(),
-  mtimeMs: z.number().nonnegative()
+  /** Whole milliseconds. The skill's 2 ms staleness window holds only if both sides floored the same
+   * nanosecond timestamp, and Node's raw `stats.mtimeMs` is fractional. */
+  mtimeMs: z.number().int().nonnegative()
 })
 
 export const SELECTION_EXCERPT_MAX_LENGTH = 2000
